Run at full screen
With the default settings, and RESIZABLE = true, I was getting crashes
on Linux. Possibly due to having a HiDPI display. The issue I
encountered is mentioned here:
https://github.com/tracyscott/RainbowStudio/blob/master/README.md#troubleshooting

But the workaround didn’t work for me. In the end it was a choice
between going full-screen or using fixed smaller dimensions.