Don’t allow pillars to be immediately toggled
Without going through all 10 of them.

We’re trying to avoid the situation where a participant picks up a head
and then immediately puts it back down, and picks it up, etc. If they
want to see the final show again we want to make sure that all heads
leave all the pillars first, before it’s possible to get back to
triggering the final show.