Allow pillars to be deactivated and reactivated
We noticed that it was possible to place some heads wrongly. So rather
than the pillar getting ‘stuck’ in its active state, we wanted to make
it so that you could remove and re-place the head elsewhere.

If the final sequence has run, this change will still not allow the
pillar to transition back to active. It will go dormant, but the
DormantState will not allow us to go back to ActiveState until all the
heads have been removed.